добавил котиков прикольных и сделал текст адекватным
