+++ internal/bot/assets/message/msg_info.txt
@@ -1,4 +1,5 @@
 Информация:
 /start - запустить бота
 /info - информация о боте
-/groupby - добавить группу
+/setgroup - добавить группу
+(использовать как: /setgroup 606-11)
