--- internal/bot/assets/message/msg_set-group.txt
@@ -1,2 +1,0 @@
-Неправильное использование команды!
-Введите команду /info и посмотрите как правильно
