+++ internal/bot/assets/message/msg_setgroup.txt
@@ -1,0 +1,2 @@
+Неправильное использование команды!
+Введите команду /info чтобы ознакомиться с правильным использованием бота
