+++ internal/bot/assets/message/msg_start.txt
@@ -1,1 +1,2 @@
-Привет, {name}. Это бот с расписанием СурГУ.\nНапиши /info, чтобы узнать больше.
+Привет, {name}. Это бот с расписанием СурГУ.
+Напиши /info, чтобы узнать больше.
